Catch syntax-error before running PHPUnit
This is required, because PHPUnit 9.6+ no longer fails for syntax errors
with exit code 255. Instead, it converts the errors to exceptions and
handles them the same as every other exception (like the ones we use in
students stubs).

Use "all-fail" to show, that exceptions are now reported as individual
failing tests (for all tests here). This was based on a syntax error,
too. But that does not work like this test expected anymore.